+++ backend/src/main/java/com/capstone/moneytree/dao/FollowsDao.java
@@ -1,4 +1,4 @@
-package com.capstone.moneytree.dao.relationshipDao;
+package com.capstone.moneytree.dao;
 
 import com.capstone.moneytree.model.relationship.Follows;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
