+++ backend/src/main/java/com/capstone/moneytree/facade/MarketInteractionsFacade.java
@@ -17,9 +17,9 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 import com.capstone.moneytree.dao.UserDao;
-import com.capstone.moneytree.dao.relationshipDao.MadeDao;
-import com.capstone.moneytree.dao.relationshipDao.OwnsDao;
-import com.capstone.moneytree.dao.relationshipDao.ToFulfillDao;
+import com.capstone.moneytree.dao.MadeDao;
+import com.capstone.moneytree.dao.OwnsDao;
+import com.capstone.moneytree.dao.ToFulfillDao;
 import com.capstone.moneytree.model.TransactionStatus;
 import com.capstone.moneytree.model.node.Stock;
 import com.capstone.moneytree.model.node.Transaction;
