+++ backend/src/main/java/com/capstone/moneytree/service/impl/DefaultTransactionService.java
@@ -18,8 +18,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.capstone.moneytree.dao.StockDao;
 import com.capstone.moneytree.dao.TransactionDao;
 import com.capstone.moneytree.dao.UserDao;
-import com.capstone.moneytree.dao.relationshipDao.MadeDao;
-import com.capstone.moneytree.dao.relationshipDao.ToFulfillDao;
+import com.capstone.moneytree.dao.MadeDao;
+import com.capstone.moneytree.dao.ToFulfillDao;
 import com.capstone.moneytree.facade.AlpacaSession;
 import com.capstone.moneytree.model.MoneyTreeOrderType;
 import com.capstone.moneytree.model.TransactionStatus;
